+++ lib/src/extensions/datetime_ext.dart
@@ -1,5 +1,5 @@
 import 'package:jiffy/jiffy.dart';
 
 extension Converting on DateTime {
-  Jiffy toJiffy() => Jiffy(this);
+  Jiffy toJiffy() => Jiffy.parseFromDateTime(this);
 }