+++ lib/src/month_item.dart
@@ -89,16 +89,20 @@ class MonthItemState extends State<MonthItem> {
     final display =
         DateTime.utc(widget.displayMonth.year, widget.displayMonth.month)
             .toJiffy();
-    _beginOffset = (widget.firstWeekDay.index > display.day - 1)
-        ? display.day - 1 + (WeekDay.values.length - widget.firstWeekDay.index)
-        : display.day - 1 - widget.firstWeekDay.index;
+    _beginOffset = (widget.firstWeekDay.index > display.dayOfWeek - 1)
+        ? display.dayOfWeek -
+            1 +
+            (WeekDay.values.length - widget.firstWeekDay.index)
+        : display.dayOfWeek - 1 - widget.firstWeekDay.index;
     _daysInMonth = display.daysInMonth;
-    _beginRange = Jiffy(Jiffy(display).subtract(days: _beginOffset));
-    _endRange = Jiffy(Jiffy(display).add(days: _daysInMonth - 1));
-    if (_endRange.day != WeekDay.sunday.index + 1) {
+    _beginRange = Jiffy.parseFromJiffy(
+        Jiffy.parseFromJiffy(display).subtract(days: _beginOffset));
+    _endRange = Jiffy.parseFromJiffy(
+        Jiffy.parseFromJiffy(display).add(days: _daysInMonth - 1));
+    if (_endRange.dayOfWeek != WeekDay.sunday.index + 1) {
       _endRange.add(
           days: WeekDay.values.length -
-              _endRange.day +
+              _endRange.dayOfWeek +
               widget.firstWeekDay.index);
     }
 
@@ -223,7 +227,7 @@ class MonthItemState extends State<MonthItem> {
     final end = _endRange;
     _weekCount = widget.forceSixWeek
         ? Contract.kMaxWeekPerMoth
-        : (end.diff(begin, Units.WEEK) + 1).toInt(); // inclusive
+        : (end.diff(begin, unit: Unit.week) + 1).toInt(); // inclusive
 
     final drawersForWeek = <List<EventProperties>>[];
     final weeks = List.generate(_weekCount, (index) {